requirements.txt
vllm>=0.11.0
torchtune>=0.6.1
torchao>= 0.11.0
tqdm>=4.67.1
pandas>=2.2.3
numpy>=2.2.6
transformers>=4.57.0
lightning>=2.6.0
safetensors>=0.5.2